feat: refactoring project structure
