+++ .eslintignore
@@ -1,3 +1,3 @@
 {
-    "ignorePatterns": ["**/src/app/core/generated/*.*"],
+    "ignorePatterns": ["**/src/app/shared/generated/*.*"]
 }
