+++ .prettierignore
@@ -1,1 +1,1 @@
-src/app/core/generated
+src/app/shared/generated
