--- src/app/core/core.module.ts
@@ -1,11 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { EntityService } from './service/entity.service';
-
-@NgModule({
-  declarations: [],
-  imports: [CommonModule],
-  exports: [],
-  providers: [EntityService],
-})
-export class CoreModule {}
