+++ src/app/shared/services/entity.service.ts
