+++ src/app/shared/shared.module.ts
@@ -1,8 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { EntityService } from './services/entity.service';
 
 @NgModule({
   declarations: [],
   imports: [CommonModule],
+  exports: [],
+  providers: [EntityService],
 })
 export class SharedModule {}
